refactor: split tabs in form builder into separate components
