+++ src/client/components/builder/LogicTab.tsx
@@ -1,0 +1,10 @@
+import React, { FC } from 'react'
+import { Container } from '@chakra-ui/react'
+
+export const LogicTab: FC = () => {
+  return (
+    <Container maxW="756px" px={0}>
+      Logic UI Components Here
+    </Container>
+  )
+}
