+++ src/client/components/builder/index.ts
@@ -5,3 +5,5 @@ export { NumericField } from './NumericField'
 export { RadioField } from './RadioField'
 export { CheckboxField } from './CheckboxField'
 export { TitleField } from './TitleField'
+export { QuestionsTab } from './QuestionsTab'
+export { LogicTab } from './LogicTab'
